+++ openfisca_core/periods/_config.py
@@ -4,11 +4,6 @@ from typing import Pattern
 
 import re
 
-DAY = "day"
-MONTH = "month"
-YEAR = "year"
-ETERNITY = "eternity"
-
 # Matches "2015", "2015-01", "2015-01-01"
 # Does not match "2015-13", "2015-12-32"
 INSTANT_PATTERN: Pattern = re.compile(r"^\d{4}(-(0[1-9]|1[012]))?(-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01]))?$")