+++ openfisca_core/periods/_units.py
@@ -2,3 +2,10 @@ DAY = "day"
 MONTH = "month"
 YEAR = "year"
 ETERNITY = "eternity"
+
+UNIT_WEIGHTS = {
+    DAY: 100,
+    MONTH: 200,
+    YEAR: 300,
+    ETERNITY: 400,
+    }